server/requirements.txt
Flask
pymongo[srv]
Flask-Cors
python-dotenv
bcrypt
gunicorn
Flask-Mail